+++ src/main/groovy/ru/vyarus/gradle/plugin/pom/PomPlugin.groovy
@@ -15,7 +15,9 @@ import org.gradle.api.plugins.WarPlugin
 import org.gradle.api.publish.PublishingExtension
 import org.gradle.api.publish.maven.MavenPublication
 import org.gradle.api.publish.maven.plugins.MavenPublishPlugin
+import org.gradle.util.GradleVersion
 import ru.vyarus.gradle.plugin.pom.xml.XmlMerger
+import ru.vyarus.gradle.plugin.pom.xml.XmlUtils
 
 /**
  * Pom plugin "fixes" maven-publish plugin pom generation: set correct scopes for dependencies. Also provides
@@ -63,6 +65,7 @@ import ru.vyarus.gradle.plugin.pom.xml.XmlMerger
  * @since 04.11.2015
  */
 @CompileStatic(TypeCheckingMode.SKIP)
+@SuppressWarnings('Println')
 class PomPlugin implements Plugin<Project> {
 
     private static final String COMPILE = 'compile'
@@ -112,7 +115,7 @@ class PomPlugin implements Plugin<Project> {
     private void activatePomModifications(Project project) {
         PublishingExtension publishing = project.publishing
         // apply to all configured maven publications (even not yet registered)
-        publishing.publications.withType(MavenPublication) { pub ->
+        publishing.publications.withType(MavenPublication) { MavenPublication pub ->
             // important to apply after possible user modifications because otherwise duplicate tags will arise
             project.afterEvaluate {
                 PomExtension extension = project.extensions.findByType(PomExtension)
@@ -128,21 +131,56 @@ class PomPlugin implements Plugin<Project> {
                         }
                     }
                 }
-                extension.configs.forEach { pom(it) }
+                applyPomModelModifiers(project, extension, pub)
                 pom.withXml {
                     Node pomXml = asNode()
+                    String before
+                    if (extension.debug) {
+                        before = XmlUtils.toString(pomXml)
+                    }
                     // do nothing for BOM
                     project.plugins.withType(JavaPlugin) {
-                        fixPomDependenciesScopes(project, extension, pomXml)
+                        fixPomDependenciesScopes(project, pub, extension, pomXml)
                     }
-                    fixDependencyManagement(extension, pomXml)
-                    applyPomModifiers(project, extension, it)
+                    fixDependencyManagement(project, pub, extension, pomXml)
+                    applyPomXmlModifiers(project, pub, extension, it)
                     if (extension.forcedVersions) {
                         validateVersions(pomXml)
+                    }
+
+                    if (extension.debug) {
+                        String after = XmlUtils.toString(pomXml)
+                        printDiff(project, extension, pub, 'Applied direct XML changes', before, after)
+                    }
+                    extension.xmlConfigsApplied = true
+                }
+            }
+        }
+    }
+
+    private void applyPomModelModifiers(Project project, PomExtension extension, MavenPublication pub) {
+        if (!extension.configs.empty) {
+            pub.pom {
+                String before
+                if (extension.debug) {
+                    // before 8.4 there were no static method for xml generation from model
+                    if (GradleVersion.current() >= GradleVersion.version('8.4')) {
+                        before = XmlUtils.toString(it)
+                    }
+                    debug(project, extension, pub, "Apply ${extension.configs.size()} pom model customizations")
+                }
+                extension.configs.each { a -> a.execute(it) }
+                if (extension.debug) {
+                    if (before) {
+                        String after = XmlUtils.toString(it)
+                        printDiff(project, extension, pub, 'Applied XML model changes', before, after)
+                    } else {
+                        println '\tPom model diff is only supported for gradle 8.4 and above'
                     }
                 }
             }
         }
+        extension.configsApplied = true
     }
 
     /**
@@ -152,7 +190,7 @@ class PomPlugin implements Plugin<Project> {
      * @param pomXml pom xml
      */
     @SuppressWarnings('MethodSize')
-    private void fixPomDependenciesScopes(Project project, PomExtension extension, Node pomXml) {
+    private void fixPomDependenciesScopes(Project project, MavenPublication pub, PomExtension extension, Node pomXml) {
         Node dependencies = pomXml.dependencies[0]
         Closure correctDependencies = { DependencySet deps, Closure action ->
             if (!dependencies || deps.empty) {
@@ -162,12 +200,24 @@ class PomPlugin implements Plugin<Project> {
             }
             dependencies.dependency.findAll {
                 deps.find { Dependency dep ->
-                    dep.group == it.groupId.text() && dep.name == it.artifactId.text()
+                    boolean res = dep.group == it.groupId.text() && dep.name == it.artifactId.text()
+                    if (res && extension.debug) {
+                        println("\t - $dep.group:$dep.name:$dep.version (original scope: ${it.scope.text()})")
+                    }
+                    res
                 }
             }.each(action)
+            if (extension.debug) {
+                println()
+            }
         }
         Closure correctScope = { DependencySet deps, String requiredScope ->
-            correctDependencies(deps) { it.scope*.value = requiredScope }
+            if (!deps.empty) {
+                debug(project, extension, pub, "Correct $requiredScope dependencies")
+            }
+            correctDependencies(deps) {
+                it.scope*.value = requiredScope
+            }
         }
 
         ConfigurationContainer configurations = project.configurations
@@ -181,8 +231,12 @@ class PomPlugin implements Plugin<Project> {
         }
 
         // OPTIONAL
-        correctDependencies(
-                configurations.optional.allDependencies - (implementation.dependencies) as DependencySet) {
+        DependencySet optionalDeps = (configurations.optional.allDependencies - implementation.dependencies)
+                as DependencySet
+        if (!optionalDeps.empty) {
+            debug(project, extension, pub, 'Correct optional dependencies')
+        }
+        correctDependencies(optionalDeps) {
             it.scope*.value = COMPILE
             it.appendNode(OPTIONAL, true.toString())
         }
@@ -200,36 +254,47 @@ class PomPlugin implements Plugin<Project> {
         }
     }
 
-    private void fixDependencyManagement(PomExtension extension, Node pomXml) {
+    private void fixDependencyManagement(Project project, MavenPublication pub, PomExtension extension, Node pomXml) {
         Node dependencyManagement = pomXml.dependencyManagement[0]
         if (!dependencyManagement) {
             return
         }
         if (extension.removedDependencyManagement) {
-            // remove dependenciesManagementSection
+            debug(project, extension, pub, 'Removing dependency management section')
             pomXml.remove(dependencyManagement)
         } else if (!extension.disabledBomsReorder) {
             // reorder BOMs (bubble BOMs to top)
             Node dependencies = dependencyManagement.dependencies[0]
-            dependencies.dependency.findAll { it.scope.text() != 'import' }.each {
-                dependencies.remove(it)
-                dependencies.append(it)
+            if (dependencies.children().size() > 1) {
+                dependencies.dependency.findAll { it.scope.text() != 'import' }.each {
+                    debug(project, extension, pub,
+                            "Move ${it.groupId.text()}:${it.artifactId.text()} import at the top")
+                    dependencies.remove(it)
+                    dependencies.append(it)
+                }
             }
         }
     }
 
-    private void applyPomModifiers(Project project, PomExtension pomExt, XmlProvider pomProvider) {
+    private void applyPomXmlModifiers(Project project, MavenPublication pub,
+                                      PomExtension pomExt, XmlProvider pomProvider) {
         Node pom = pomProvider.asNode()
-        pomExt.rawConfigs.each {
-            XmlMerger.mergePom(pom, it)
+        if (!pomExt.rawConfigs.empty) {
+            debug(project, pomExt, pub, "Apply ${pomExt.rawConfigs.size()} withXml closures")
         }
+        pomExt.rawConfigs.each { XmlMerger.mergePom(pom, it) }
 
+        if (!pomExt.xmlModifiers.empty) {
+            debug(project, pomExt, pub, "Apply ${pomExt.xmlModifiers.size()} withPomXml customizations")
+        }
         pomExt.xmlModifiers.each { it.execute(pomProvider) }
         // apply defaults if required
         if (!pom.name) {
+            debug(project, pomExt, pub, 'Apply default name')
             pom.appendNode('name', project.name)
         }
         if (project.description && !pom.description) {
+            debug(project, pomExt, pub, 'Apply default description')
             pom.appendNode('description', project.description)
         }
     }
@@ -250,6 +315,27 @@ class PomPlugin implements Plugin<Project> {
             throw new GradleException('No versions resolved for the following dependencies. Most likely, there are ' +
                     'no required repositories declared. Declare missed repositories with, for example: ' +
                     'repositories { mavenCentral() }\n' + errors.join(''))
+        }
+    }
+
+    private void debug(Project project, PomExtension ext, MavenPublication pub, String message) {
+        if (ext.debug) {
+            String prj = project == project.rootProject ? '' : " (in ${XmlUtils.CYAN}$project.name${XmlUtils.RESET})"
+            println "POM> $message for ${XmlUtils.PURPLE}$pub.name${XmlUtils.RESET} publication$prj"
+        }
+    }
+
+    @SuppressWarnings('ParameterCount')
+    private void printDiff(Project project, PomExtension extension, MavenPublication pub,
+                           String message, String before, String after) {
+        String diff = XmlUtils.diffShifted(before, after, '\t')
+        if (diff.empty) {
+            debug(project, extension, pub, 'No xml modifications detected')
+        } else {
+            println()
+            debug(project, extension, pub, "--------------------------------- $message")
+            println diff
+            println '     --------------------------------'
         }
     }
 }