+++ src/test/groovy/ru/vyarus/gradle/plugin/pom/LegacyKitTest.groovy
@@ -43,12 +43,15 @@ class LegacyKitTest extends AbstractKitTest {
                 }
             }
 
-            maven.pom {
-                developers {
-                    developer {
-                        id = "dev"
-                        name = "Dev Dev"
-                        email = "dev@gmail.com"
+            maven {
+                debug()
+                pom {
+                    developers {
+                        developer {
+                            id = "dev"
+                            name = "Dev Dev"
+                            email = "dev@gmail.com"
+                        }
                     }
                 }
             }