+++ src/test/groovy/ru/vyarus/gradle/plugin/pom/UpstreamKitTest.groovy
@@ -41,12 +41,15 @@ class UpstreamKitTest extends AbstractKitTest {
                 }
             }
 
-            maven.pom {
-                developers {
-                    developer {
-                        id = "dev"
-                        name = "Dev Dev"
-                        email = "dev@gmail.com"
+            maven {
+                debug()
+                pom {
+                    developers {
+                        developer {
+                            id = "dev"
+                            name = "Dev Dev"
+                            email = "dev@gmail.com"
+                        }
                     }
                 }
             }